increase timeout to account for length of page
From Applitools:
Lazy load scrolls 300 pixels, and waits 2 seconds before continuing the scroll. This will continue until we reach the end of the page. As an example, when scrolling the applitools.com page, it has a height of 12,500 pixels.
This means that lazy load will take 84 seconds to complete.
Since timeout is set to 60 seconds, Playwright will timeout the tests before the lazy load can complete.